+++ vmdhelper.h
@@ -5,7 +5,7 @@ class VMDHelper {
 public:
 	VMDHelper(std::string pdb_file, std::string amber_topology_file, std::string namd2_path, std::string vmd_path);
 
-	float calculateEnergy();
+	double calculateEnergy();
 
 private:
